+++ test/db/postgresql/change_column_test.rb
@@ -1,6 +1,0 @@
-require 'db/postgres'
-require 'change_column_test_methods'
-
-class PostgreSQLChangeColumnTest < Test::Unit::TestCase
-  include ChangeColumnTestMethods
-end